@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2017, Sensirion AG
+ * Copyright (c) 2024, Draekko RAND
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -30,7 +31,7 @@
  */
 package com.sensirion.smartgadget.peripheral.rht_sensor;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.sensirion.smartgadget.peripheral.PeripheralConnectionStateListener;
 
