@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2017, Sensirion AG
+ * Copyright (c) 2024, Draekko RAND
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -30,7 +31,7 @@
  */
 package com.sensirion.smartgadget.peripheral.rht_utils;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 
 public abstract class TemperatureConverter {
